+++ src/main/java/com/springapp/mvc/repository/FieldOfficerRepository.java
@@ -4,5 +4,5 @@ import com.springapp.mvc.entity.FieldOfficer;
 import com.springapp.mvc.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FieldOfficerRepository extends JpaRepository<FieldOfficer, Long> {
+public interface FieldOfficerRepository extends JpaRepository<FieldOfficer, Integer> {
 }
