+++ src/main/java/com/springapp/mvc/repository/LoanRepository.java
@@ -4,5 +4,5 @@ import com.springapp.mvc.entity.Loan;
 import com.springapp.mvc.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<Loan, Long> {
+public interface LoanRepository extends JpaRepository<Loan, Integer> {
 }
