+++ src/main/java/com/springapp/mvc/repository/TransferDetailRepository.java
@@ -1,8 +1,10 @@
 package com.springapp.mvc.repository;
 
-import com.springapp.mvc.entity.Bank;
 import com.springapp.mvc.entity.TransferDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransferDetailRepository extends JpaRepository<TransferDetail, Long> {
+/**
+ * Created by ssvh on 7/24/16.
+ */
+public interface TransferDetailRepository extends JpaRepository<TransferDetail,Integer>{
 }
